requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
clickhouse-connect==0.7.19
SQLAlchemy>=2.0
psycopg[binary]>=3.1
bcrypt==5.0.0